rename setting examples for ESPboy
